// interfaces/IFriend.ts
export interface IFriend {
  id: string;
  age: number;
  ageStatus: string;
  nameUA: string;
  name: string;
  shortDescriptionUA: string;
  shortDescription: string;
  descriptionUA: string;
  description: string;
  categoryUA: string;
  category: string;
  price: number;
  sexUA: string;
  sex: string;
  wc: boolean;
  images: string[];
}